Replace Cloudwatch decorator with driver
